client protocol + factory
