+++ Sources/Vapor/Engine/Client/ClientProtocol.swift
@@ -1,0 +1,12 @@
+import HTTP
+import Transport
+import URI
+import TLS
+
+public protocol ClientProtocol: Responder {
+    init(
+        hostname: String,
+        port: Port,
+        _ securityLayer: SecurityLayer
+    ) throws
+}
