+++ Sources/Vapor/Engine/Client/Responder+Helpers.swift
@@ -1,4 +1,50 @@
 import HTTP
+import URI
+
+// MARK: Base
+
+extension Responder {
+    /// Creates a new client from the information in the
+    /// Request URI and uses it to respond to the request.
+    public func respond(
+        to req: Request,
+        through middleware: [Middleware] = []
+    ) throws -> Response {
+        return try middleware
+            .chain(to: self)
+            .respond(to: req)
+    }
+    
+    /// Creates a new client and calls `.respond()`
+    /// using the request method and uri provided.
+    public func request(
+        _ method: Method,
+        _ uri: String,
+        query: [String: NodeRepresentable] = [:],
+        _ headers: [HeaderKey: String] = [:],
+        _ body: BodyRepresentable? = nil,
+        through middleware: [Middleware] = []
+        ) throws  -> Response {
+        var uri = try URI(uri)
+        
+        var q: [String: CustomStringConvertible] = [:]
+        try query.forEach { key, value in
+            q[key] = try value.makeNode(in: nil).string ?? ""
+        }
+        uri.append(query: q)
+        
+        let req = Request(method: method, uri: uri)
+        req.headers = headers
+        
+        
+        if let body = body {
+            req.body = body.makeBody()
+        }
+        return try respond(to: req, through: middleware)
+    }
+}
+
+// MARK: Method specific
 
 extension Responder {
     /// Calls `.request(.get, ...)`
