+++ Sources/Vapor/Engine/SecurityLayer.swift
@@ -5,3 +5,10 @@ public enum SecurityLayer {
     case none
     case tls(TLS.Context)
 }
+
+
+extension String {
+    var port: Port {
+        return isSecure ? 443 : 80
+    }
+}
